controllers/cgroup: add mount option "cpuacct" in test03
"cpu,cpuacct" will be mounted together in most distribution,
AGL system also included.
But it's tough to umount "cpu,cpuacct" on Rcar-M3, which lead to
mount "cpu" controller alone will be fail in this case.

Maybe the ultimate cause is Rcar-M3 arch(A53 not enabled), so
this just a temporary solution.

Signed-off-by: Liu Wenlong <liuwl.fnst@cn.fujitsu.com>