net.nfs/nfslock*: reduce the test pressure
reduce the line of tmp file from 16384 to 100, because this is
very slow test(it need about 15s to change 1 line on M3&DB)

The line is 16348 before and don't know the meaning of 16384 now.
So this just a temporary solution and maybe the value shoule be
configurable in the future.

Target test machine:
* M3: Rcar-M3
* DB: Dragonboard-410c

Signed-off-by: Liu Wenlong <liuwl.fnst@cn.fujitsu.com>