net.tcp_cmds/xinetd: change the retrieval keywords and the way to quit "telnet".
The expect output by test program is not same with before.
The contents of output were changed a lot, so change the
matching keywords to compatible with multiple outputs.

Changed "echo '^D' | telnet $IP" to "sleep 2 | telnet $IP",
because "echo '^D' runs too fast to get the desired message
for Rcar-M3 and dragonboard-410c.

Signed-off-by: Liu Wenlong <liuwl.fnst@cn.fujitsu.com>