net.nfs/nfslock*: reduce the test pressure.
reduce the line of tmp file from 16384 to 100, because this is a
very slow test(it need about 15s to change 1 line on  Rcar-M3 and dragonboard-410c)
The needed time before: about several days;
The needed time now: about half an hour.

The line is 16348 before.
This just a temporary solution and may make the value more configurable in the future.

Signed-off-by: Liu Wenlong <liuwl.fnst@cn.fujitsu.com>